requirements.txt
streamlit
python-dotenv
openai
azure-search-documents
pypdf
tiktoken
